requirements.txt
paho-mqtt==1.6.1
adafruit-circuitpython-dht==4.0.4
RPi.GPIO==0.7.1
picamera2==0.3.18
opencv-python==4.10.0.84
python-dotenv==1.0.1
requests==2.32.3
